Extracted offline classes because CodeClimate
